requirements.txt
Flask
Flask-Login
Werkzeug
google-genai
python-dotenv
sqlalchemy
flask_sqlalchemy
google.generativeai

